Prevent location decimals from mapping to wrong index.
Mapping indexes to embeddings works by multiplying y emb lenth by
y location and then adding x location.
EX:
point (14, 72)
Will map to the flattened embedding by 14 * 168 + 72=2424

The issue is when the location is a decimal
EX:
point (14.78, 72)
14.78 * 168 + 72 = 2555.04 was gets rounded to 2555
This is outside of y dim 14's "row" so 14 * 168 + 168 = 2520

This can also cause a runtime error if the final row has a point.
EX:
point (167.78, 72)
167.78 * 168 + 72 = 28259.04 which gets rounded down to 28259
This is out of bound for the flattened emb which is size 168 * 168 = 28224