Fix normalization, numerical stability in loss.
The dim param in normalization was changed  because this apreaded to be what was working,
on closer analysis this was incorrect. The pameter that has been specified (and recently changed from 1 to 2) was actually the p var in norm. So the correct dim was specified (1)
and was given to the correct parameter.

The other issue that appreared at lower temperature params, was numerical instabilty
that was partly caused by just subtracting exp(1 / t) from the sumed simularities.
Instead the diagonal is not counted in the summation fixing this issue.

The loss was also changed to be a proper torch Module to precomputer the diag matrix for
the summation change.