Add sparse models and training.
SparseSegHead:
	Can train when give a list of embeddings or a spatial embedding with points.
	Can transition linear layers to conv layers by overwritting the `eval()` method.

Made two new pipelines:
	* sparse_sh - trains the sparse segheads on pre-made embeddings
	*sparse_baseline - trains end to end baselines with the sparse_sh

The pipelines require a new GP node `PointsLabelsSource` that can take a list of points
and labels and return the randomly chosen points either into a point Array or Graph. Can
put labels in Array if requested.

These require a new loss to deal with the points. A PointLoss is introduced that can handle
a list of predictions and spatial gt OR a list of predictions and a list of gt.

To train the non-baseline sparse models:

	1. Run generate_points.py that will grab random points. Currently is a dirty script
	and has hard coded paths. Will be changed in the future.
	2. Train a contrastive model
	3. Run the train.py in the new `emb` mode to generate emb for each contrastive model
	and checkpoint.
	4. Run the segmentation model